+++ lib/onlyofficePlugin.php
@@ -24,18 +24,17 @@
  */
 class OnlyofficePlugin extends Plugin implements HookPluginInterface
 {
+
+    /**
+     * OnlyofficePlugin name.
+     */
+    private $pluginName = "onlyoffice";
+
     /**
      * OnlyofficePlugin constructor.
      */
     protected function __construct()
     {
-        $bannerTemplate = new Template('ONLYOFFICE Docs Cloud');
-        $bannerTemplate->assign('docs_cloud_link', AppConfig::GetLinkToDocs());
-        $bannerTemplate->assign('banner_title', $this->get_lang('DocsCloudBannerTitle'));
-        $bannerTemplate->assign('banner_main_text', $this->get_lang('DocsCloudBannerMain'));
-        $bannerTemplate->assign('banner_button_text', $this->get_lang('DocsCloudBannerButton'));
-        $docsCloudBanner = $bannerTemplate->fetch('/onlyoffice/layout/get_docs_cloud_banner.tpl');
-
         parent::__construct(
             "1.2.0",
             "Asensio System SIA",
@@ -43,7 +42,8 @@ class OnlyofficePlugin extends Plugin implements HookPluginInterface
                 "enable_onlyoffice_plugin" => "boolean",
                 "document_server_url" => "text",
                 "jwt_secret" => "text",
-                $docsCloudBanner => "html",
+                "jwt_header" => "text",
+                "document_server_internal" => "text",
             ]
         );
     }
@@ -102,5 +102,183 @@ class OnlyofficePlugin extends Plugin implements HookPluginInterface
 
         $viewObserver = OnlyofficeItemViewObserver::create();
         HookDocumentItemView::create()->detach($viewObserver);
+    }
+
+    /**
+     * Get status of demo server
+     *
+     * @return bool
+     */
+    public function useDemo() {
+        return $this->getDemoData()["enabled"] === true;
+    }
+
+    /**
+     * Get demo data
+     *
+     * @return array
+     */
+    public function getDemoData() {
+        $data = api_get_setting('onlyoffice_connect_demo_data')[0];
+
+        if (empty($data)) {
+            $data = [
+                "available" => true,
+                "enabled" => false
+            ];
+            api_add_setting(json_encode($data), 'onlyoffice_connect_demo_data', null, 'setting', 'Plugins');
+            return $data;
+        }
+        $data = json_decode($data, true);
+
+        if (isset($data['start'])) {
+            $overdue = $data["start"];
+            $overdue += 24*60*60*AppConfig::GetDemoParams()["TRIAL"];
+            if ($overdue > time()) {
+                $data["available"] = true;
+                $data["enabled"] = $data["enabled"] === true;
+            } else {
+                $data["available"] = false;
+                $data["enabled"] = false;
+            }
+            api_set_setting('onlyoffice_connect_demo_data', json_encode($data));
+        }
+        return $data;
+    }
+
+    /**
+     * Switch on demo server
+     *
+     * @param bool $value - select demo
+     *
+     * @return bool
+     */
+    public function selectDemo($value) {
+        $data = $this->getDemoData();
+
+        if ($value === true && !$data["available"]) {
+            return false;
+        }
+
+        $data["enabled"] = $value === true;
+
+        if (!isset($data["start"])) {
+            $data["start"] = time();
+        }
+        api_set_setting('onlyoffice_connect_demo_data', json_encode($data));
+        return true;
+    }
+
+    /**
+     * Get the document server url
+     *
+     * @param bool $origin - take origin
+     *
+     * @return string
+     */
+    public function getDocumentServerUrl($origin = false) 
+    {
+        if (!$origin && $this->useDemo()) {
+            return AppConfig::GetDemoParams()["ADDR"];
+        }
+
+        $url = $this->get("document_server_url");
+        if ($url !== null && $url !== "/") {
+            $url = rtrim($url, "/");
+            if (strlen($url) > 0) {
+                $url = $url . "/";
+            }
+        }
+        return $url;
+    }
+
+    /**
+     * Get the document service address available from Chamilo from the application configuration
+     *
+     * @param bool $origin - take origin
+     *
+     * @return string
+     */
+    public function getDocumentServerInternalUrl($origin = false) {
+        if (!$origin && $this->useDemo()) {
+            return $this->getDocumentServerUrl();
+        }
+
+        $url = api_get_setting('onlyoffice_document_server_internal')[$this->pluginName];
+        if (empty($url)) {
+            $url = AppConfig::InternalUrl();
+        }
+        if (!$origin && empty($url)) {
+            $url = $this->getDocumentServerUrl();
+        }
+        return $url;
+    }
+
+    /**
+     * Replace domain in document server url with internal address from configuration
+     *
+     * @param string $url - document server url
+     *
+     * @return string
+     */
+    public function replaceDocumentServerUrlToInternal($url) {
+        $documentServerUrl = $this->getDocumentServerInternalUrl();
+        if (!empty($documentServerUrl)) {
+            $from = $this->getDocumentServerUrl();
+
+            if (!preg_match("/^https?:\/\//i", $from)) {
+                $parsedUrl = parse_url($url);
+                $from = $parsedUrl["scheme"] . "://" . $parsedUrl["host"] . (array_key_exists("port", $parsedUrl) ? (":" . $parsedUrl["port"]) : "") . $from;
+            }
+
+            if ($from !== $documentServerUrl)
+            {
+                $url = str_replace($from, $documentServerUrl, $url);
+            }
+        }
+
+        return $url;
+    }
+
+    /**
+     * Get the document service secret key from the application configuration
+     *
+     * @param bool $origin - take origin
+     *
+     * @return string
+     */
+    public function getDocumentServerSecret($origin = false) {
+        if (!$origin && $this->useDemo()) {
+            return AppConfig::GetDemoParams()["SECRET"];
+        }
+        return $this->get("jwt_secret");
+    }
+
+    /**
+     * Get the jwt header setting
+     *
+     * @param bool $origin - take origin
+     *
+     * @return string
+     */
+    public function getJwtHeader($origin = false) {
+        if (!$origin && $this->useDemo()) {
+            return AppConfig::GetDemoParams()["HEADER"];
+        }
+
+        $header = api_get_setting('onlyoffice_jwt_header')["onlyoffice"];
+        if (empty($header)) {
+            $header = AppConfig::JwtHeader() ? AppConfig::JwtHeader() : "Authorization";
+        }
+        return $header;
+    }
+
+    /**
+     * Get link to plugin settings
+     *
+     * @return string
+     */
+    public function getConfigLink() {
+        return api_get_path(WEB_PATH)."main/admin/configure_plugin.php?name=".$this->pluginName;
     }
 }